added CryptoWallet

+++ src/main/java/ra/common/currency/Coin.java
@@ -2,8 +2,10 @@ package ra.common.currency;
 
 import ra.common.JSONSerializable;
 
+import java.math.BigInteger;
+
 public interface Coin extends JSONSerializable {
     String symbol();
     String name();
-    Double value();
+    BigInteger value();
 }
